+++ src/main/java/io/github/astrapi69/swing/panel/desktoppane/JDesktopPanePanel.java
@@ -1,0 +1,135 @@
+/**
+ * The MIT License
+ *
+ * Copyright (C) 2021 Asterios Raptis
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to
+ * the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+ * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+package io.github.astrapi69.swing.panel.desktoppane;
+
+import java.awt.BorderLayout;
+import java.awt.Component;
+
+import javax.swing.JDesktopPane;
+import javax.swing.JInternalFrame;
+
+import io.github.astrapi69.swing.component.factory.JComponentFactory;
+import io.github.astrapi69.swing.desktoppane.SingletonDesktopPane;
+import io.github.astrapi69.swing.utils.JInternalFrameExtensions;
+import lombok.Getter;
+
+import io.github.astrapi69.model.api.IModel;
+import io.github.astrapi69.swing.base.BasePanel;
+import lombok.Setter;
+
+/**
+ * Panel with a JDesktopPane
+ */
+@Getter
+public class JDesktopPanePanel<T> extends BasePanel<T>
+{
+
+	/** The Constant serialVersionUID. */
+	private static final long serialVersionUID = 1L;
+
+	/** The current visible internal frame. */
+	@Setter
+	JInternalFrame currentVisibleInternalFrame;
+
+	/** The {@link JDesktopPane} */
+	JDesktopPane desktopPane;
+
+	/**
+	 * Instantiates a new new {@link JDesktopPanePanel} object
+	 */
+	public JDesktopPanePanel()
+	{
+	}
+
+	/**
+	 * Instantiates a new new {@link JDesktopPanePanel} object
+	 *
+	 * @param model
+	 *            the model
+	 */
+	public JDesktopPanePanel(final IModel<T> model)
+	{
+		super(model);
+	}
+
+	/**
+	 * Factory method for creating the new {@link JDesktopPane}. This method is invoked in the
+	 * constructor from the derived classes and can be overridden so users can provide their own
+	 * version of a {@link JDesktopPane}
+	 *
+	 * @return the new {@link JDesktopPane}
+	 */
+	protected JDesktopPane newJDesktopPane()
+	{
+		return SingletonDesktopPane.getInstance();
+	}
+
+	/**
+	 * {@inheritDoc}
+	 */
+	@Override
+	protected void onInitializeComponents()
+	{
+		super.onInitializeComponents();
+
+		desktopPane = newJDesktopPane();
+		desktopPane.setName("JDesktopPanePanel.desktopPane");
+	}
+
+	/**
+	 * {@inheritDoc}
+	 */
+	@Override
+	protected void onInitializeLayout()
+	{
+		super.onInitializeLayout();
+		setLayout(new BorderLayout());
+		add(desktopPane, BorderLayout.CENTER);
+	}
+
+	/**
+	 * Replace the current internal frame with a new internal frame with the given {@link Component}
+	 * as content.
+	 *
+	 * @param title
+	 *            the title
+	 * @param component
+	 *            the component
+	 */
+	public void replaceInternalFrame(final String title, final Component component)
+	{
+		if (getCurrentVisibleInternalFrame() != null)
+		{
+			getCurrentVisibleInternalFrame().dispose();
+		}
+		// create internal frame
+		final JInternalFrame internalFrame = JComponentFactory.newInternalFrame(title, true, true,
+			true, true);
+		JInternalFrameExtensions.addComponentToFrame(internalFrame, component);
+		JInternalFrameExtensions.addJInternalFrame(desktopPane, internalFrame);
+		setCurrentVisibleInternalFrame(internalFrame);
+	}
+
+}